requirements_qwen2_5_vl.txt
absl-py==2.3.1
accelerate==1.11.0
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
arxiv==2.3.0
asttokens==3.0.0
async-lru==2.0.5
async-timeout==5.0.1
attrdict==2.0.1
attrs==25.4.0
av==16.0.1
babel==2.17.0
beautifulsoup4==4.14.2
bitsandbytes==0.45.0
bleach==6.3.0
blinker==1.9.0
blis==1.3.0
brotli==1.2.0
cachetools==6.2.1
catalogue==2.0.10
certifi==2022.12.7
cffi==2.0.0
charset-normalizer==2.1.1
click==8.3.0
cloudpathlib==0.23.0
colorama==0.4.6
comm==0.2.3
confection==0.1.5
contourpy==1.3.2
crcmod==1.7
cryptography==46.0.3
cycler==0.12.1
cymem==2.0.11
datasets==4.4.1
debugpy==1.8.17
decorator==5.2.1
deepspeed==0.18.2
defusedxml==0.7.1
dill==0.4.0
distro==1.9.0
duckduckgo_search==5.3.1b1
einops==0.8.1
et_xmlfile==2.0.0
exceptiongroup==1.3.0
executing==2.2.1
fastjsonschema==2.21.2
feedparser==6.0.12
filelock==3.14.0
fonttools==4.60.1
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2025.9.0
ftfy==6.3.1
func_timeout==4.3.5
gitdb==4.0.12
GitPython==3.1.45
griffe==0.49.0
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hjson==3.1.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
hyperframe==6.1.0
idna==3.4
ImageIO==2.37.2
ipykernel==7.1.0
ipython==8.37.0
ipywidgets==8.1.8
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.1
jmespath==0.10.0
joblib==1.5.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.10
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
kiwisolver==1.4.9
lagent==0.2.4
langcodes==3.5.0
language_data==1.3.0
lark==1.3.1
lazy_loader==0.4
loguru==0.7.3
marisa-trie==1.3.1
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.7
matplotlib-inline==0.2.1
mdurl==0.1.2
mistune==3.1.4
mmcv==2.2.0
-e git+https://github.com/open-mmlab/mmdetection.git@cfd5d3a985b0249de009b67d04f37263e11cdf3d#egg=mmdet
mmengine==0.10.4
-e git+https://github.com/open-mmlab/mmsegmentation.git@b040e147adfa027bbc071b624bedf0ae84dfc922#egg=mmsegmentation
model-index==0.1.11
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
multiprocess==0.70.18
murmurhash==1.0.13
narwhals==2.10.2
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
ninja==1.13.0
notebook==7.4.7
notebook_shim==0.2.4
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.7.0.84
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.1.0.106
nvidia-cusparselt-cu12==0.7.1
nvidia-ml-py==13.580.82
nvidia-nccl-cu11==2.19.3
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.1.105
open_clip_torch==3.2.0
openai==2.7.1
opencv-python==4.8.1.78
opendatalab==0.0.10
openmim==0.3.9
openpyxl==3.1.5
openxlab==0.1.3
ordered-set==4.1.0
oss2==2.17.0
overrides==7.7.0
packaging==24.2
pandas==2.3.3
pandocfilters==1.5.1
panopticapi @ git+https://github.com/cocodataset/panopticapi.git@7bb4655548f98f3fedc07bf37e9040a992b054b0
parso==0.8.5
peft==0.17.1
pexpect==4.9.0
phx-class-registry==4.1.0
pillow==11.3.0
platformdirs==4.5.0
preshed==3.0.10
prettytable==3.16.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.0
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==21.0.0
pycocotools==2.0.10
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.4
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-json-logger==4.0.0
pytz==2023.4
PyYAML==6.0.3
pyzmq==27.1.0
qwen-vl-utils==0.0.14
referencing==0.37.0
regex==2025.11.3
requests==2.28.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==13.4.2
rpds-py==0.28.0
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
Send2Trash==1.8.3
sentencepiece==0.2.1
sgmllib3k==1.0.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
smart_open==7.4.4
smmap==5.0.2
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.8
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
stack-data==0.6.3
streamlit==1.51.0
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
termcolor==3.2.0
terminado==0.18.1
terminaltables==3.1.10
thinc==8.3.7
threadpoolctl==3.6.0
tifffile==2025.5.10
tiktoken==0.12.0
timeout-decorator==0.5.0
timm==1.0.22
tinycss2==1.4.0
tokenizers==0.21.4
toml==0.10.2
tomli==2.3.0
torch==2.2.2+cu118
torchvision==0.17.2+cu118
tornado==6.5.2
tqdm==4.65.2
traitlets==5.14.3
transformers==4.51.3
transformers-stream-generator==0.0.5
triton==2.2.0
typer==0.20.0
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uri-template==1.3.0
urllib3==1.26.13
wasabi==1.1.3
watchdog==6.0.0
wcwidth==0.2.14
weasel==0.4.2
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
Werkzeug==3.1.3
widgetsnbextension==4.0.15
wrapt==2.0.0
xtuner==0.1.23
xxhash==3.6.0
yapf==0.43.0
yarl==1.22.0
